+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ### Unreleased
 
+* Eagerly initialize `Model#cells`
+
 ### 0.5.1 (2016-12-01)
 
 * Convert shardable value to integer for index tables